Implement CLEAN architecture on (Domain | Servives)
