+++ LibraryApp/LibraryApp.Api/Program.cs
@@ -1,4 +1,4 @@
-using LibraryApp.Console.Services;
+using LibraryApp.Services;
 
 var builder = WebApplication.CreateBuilder(args);
