+++ LibraryApp/LibraryApp.Domain/Book.cs
@@ -1,0 +1,24 @@
+﻿namespace LibraryApp.Domain
+{
+    public sealed class Book : LibraryItem
+    {
+        public string Author { get; }
+        public int Pages { get; }
+
+        public Book(int id, string title, string author) : this(id, title, author, pages: 0) { }
+
+        public Book(int id, string title, string author, int pages) : base(id, title)
+        {
+            Author = string.IsNullOrWhiteSpace(title) ? "Unknown" : author.Trim();
+            Pages = pages < 0 ? 0 : pages;
+        }
+
+        public override string GetInfo()
+        {
+            return $"Book [Id={Id}], Title={Title}, Author={Author}, Pages={Pages}, IsBorrowed={IsBorrowed}";
+        }
+    }
+
+    public sealed record BookDto(string? Title, string? Author, int Pages);
+
+}
