+++ LibraryApp/LibraryApp.Domain/Magazine.cs
@@ -1,4 +1,4 @@
-﻿namespace LibraryApp.Console.Domain
+﻿namespace LibraryApp.Domain
 {
     public sealed class Magazine : LibraryItem
     {
