+++ LibraryApp/LibraryApp.Services/LibraryService.cs
@@ -1,6 +1,6 @@
-﻿using LibraryApp.Console.Domain;
+﻿using LibraryApp.Domain;
 
-namespace LibraryApp.Console.Services
+namespace LibraryApp.Services
 {
     public sealed class LibraryService : ILibraryService
     {
